+++ miso-service/src/main/java/uk/ac/bbsrc/tgac/miso/service/SubmissionService.java
@@ -2,15 +2,12 @@ package uk.ac.bbsrc.tgac.miso.service;
 
 import java.io.IOException;
 import java.util.Collection;
-import java.util.Map;
 
 import uk.ac.bbsrc.tgac.miso.core.data.Submission;
+import uk.ac.bbsrc.tgac.miso.core.service.SaveService;
 
-public interface SubmissionService {
-  Submission get(long id) throws IOException;
+public interface SubmissionService extends SaveService<Submission> {
 
   Collection<Submission> list() throws IOException;
-  long save(Submission submission) throws IOException;
 
-  Map<String, Integer> getColumnSizes() throws IOException;
 }